Refactor event show page render method more simply
Move ParticipateButton to buttons directory.
